pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "Alpha's heartbeat - scheduled job runner"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "redis>=5.0",
    "icalendar>=5.0",
    "pendulum>=3.0",
    "watchfiles>=1.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "psycopg[binary]>=3.1",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "opentelemetry-instrumentation-logging>=0.44b0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pulse"]
